+++ Providers/Resgrid.Providers.Number/Models/BulkVSMessageResponse.cs
@@ -1,0 +1,9 @@
+namespace Resgrid.Providers.NumberProvider.Models
+{
+	internal class BulkVSMessageResponse
+	{
+		public string MessageId { get; set; }
+		public string Status    { get; set; }
+		public string Error     { get; set; }
+	}
+}